# parser: parse gbk encoding SQL token by token instead of the whole SQL