--- a/parser/charset/encoding_gbk.go
+++ b/parser/charset/encoding_gbk.go
@@ -55,6 +55,20 @@ func (e *encodingGBK) Peek(src []byte) []byte {
 	return src
 }
 
+func (e *encodingGBK) MbLen(bs string) int {
+	if len(bs) < 2 {
+		return 0
+	}
+
+	if 0x81 <= bs[0] && bs[0] <= 0xf4 {
+		if (0x40 <= bs[1] && bs[1] <= 0x7e) || (0x80 <= bs[1] && bs[1] <= 0xfe) {
+			return 2
+		}
+	}
+
+	return 0
+}
+
 // ToUpper implements Encoding interface.
 func (e *encodingGBK) ToUpper(d string) string {
 	return strings.ToUpperSpecial(GBKCase, d)