--- a/parser/charset/encoding_gbk.go
+++ b/parser/charset/encoding_gbk.go
@@ -60,7 +60,7 @@ func (e *encodingGBK) MbLen(bs string) int {
 		return 0
 	}
 
-	if 0x81 <= bs[0] && bs[0] <= 0xf4 {
+	if 0x81 <= bs[0] && bs[0] <= 0xfe {
 		if (0x40 <= bs[1] && bs[1] <= 0x7e) || (0x80 <= bs[1] && bs[1] <= 0xfe) {
 			return 2
 		}