… br/pkg/lightning/restore/tidb_test.go
… executor/partition_table_test.go
… parser/ast/expressions_test.go
… parser/digester_test.go
… parser/lexer_test.go
… parser/misc.go
… parser/parser_test.go
… parser/yy_parser.go