<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close https://github.com/pingcap/tidb/issues/30765
close https://github.com/pingcap/tidb/issues/29382

this pr is based on https://github.com/pingcap/tidb/pull/29573. please only review the second commit.
It will make the parser compatible with MySQL in parsing gbk encoding characters.

```go
// benchcode
func BenchmarkParser(b *testing.B) {
	p := parser.New()
	sql := "select '" + strings.Repeat("这是一个测试", 10000) + "' as " + strings.Repeat("aaaaa", 10000)
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		_, _, err := p.ParseSQL(sql)
		require.NoError(b, err)
	}
}
```

`client=utf8 & connection=utf8`
```
before                       after
876	   1410416 ns/op     775	   1497224 ns/op
829	   1427232 ns/op     831	   1486620 ns/op
686	   1570810 ns/op     660	   1678825 ns/op
747	   1512891 ns/op     622	   1718708 ns/op
781	   1512988 ns/op     706	   1662557 ns/op
```
![image](https://user-images.githubusercontent.com/20839912/147526848-c9e2c5ec-42f9-4cf4-be86-0e9fc3f3e1fd.png)

`client=gbk & connection=utf8`
`client=gbk & connection=gbk`
```
240	   5076879 ns/op
237	   5216370 ns/op
232	   5058759 ns/op
234	   5134991 ns/op
234	   5142602 ns/op
```
![image](https://user-images.githubusercontent.com/20839912/147526757-d1fb8f4c-edf7-4cfa-a262-05518ed217a4.png)

we need optimize `convert2Connection` function
### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```